% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeReps.R
\name{analyzeReps}
\alias{analyzeReps}
\title{Analyze Replicates}
\usage{
analyzeReps(analysis, params, num.cores = 1)
}
\arguments{
\item{analysis}{character vector of EBV analysis to run (`ldNe`, `g2`, 
`het`, or `froh`).}

\item{params}{parameter list output from \code{runEBVsim()}.}

\item{num.cores}{number of cores to use to split replicate analyses among.}
}
\value{
data frame of summary metrics
}
\description{
Conduct analysis of all replicates for a given EBV
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
