% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availRuns.R
\name{availRuns}
\alias{availRuns}
\title{Available Runs}
\usage{
availRuns(drive.id)
}
\arguments{
\item{drive.id}{ID of Google Drive where runs are stored}
}
\value{
character vector of labels of runs available for download on the
  Google Drive requested in 'drive.id'
}
\description{
Return run labels available on a Google Drive.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
