% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_froh.R
\name{calc_froh}
\alias{calc_froh}
\title{Calculate Runs of Homozygosity}
\usage{
calc_froh(x)
}
\arguments{
\item{x}{data frame of genotypes as read from simulation replicate .csv file.}
}
\value{
data frame of ROH summary
}
\description{
Calculate Runs of Homozygosity
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
