% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_g2.R
\name{calc_g2}
\alias{calc_g2}
\title{Calculate inbreeding (g2)}
\usage{
calc_g2(x, nboot = 100)
}
\arguments{
\item{x}{data frame of genotypes as read from simulation replicate .csv file.}

\item{nboot}{number of bootstrap replicates to estimate confidence interval.}
}
\value{
data frame of g2 and confidence intervals for each population
}
\description{
Calculate inbreeding (g2)
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
