% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_heterozygosity.R
\name{calc_heterozygosity}
\alias{calc_heterozygosity}
\title{Calculate Heterozygosity}
\usage{
calc_heterozygosity(x)
}
\arguments{
\item{x}{data frame of genotypes as read from simulation replicate .csv file}
}
\value{
data frame of mean, median and 95% CI of observed and expected 
  heterozygosity for each population
}
\description{
Calculate observed and expected heterozygosity
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
