% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ldNe.R
\name{calc_ldNe}
\alias{calc_ldNe}
\title{Calculate ldNe}
\usage{
calc_ldNe(x)
}
\arguments{
\item{x}{data frame of genotypes as read from simulation replicate .csv file}
}
\value{
estimated Ne for each stratum in file with lower and upper CIs
}
\description{
Calculate ldNe for data frame of genotypes
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
