% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFreqs.R
\name{calcFreqs}
\alias{calcFreqs}
\alias{alleleProp}
\title{Calculate Allele Frequencies}
\usage{
calcFreqs(snps)

alleleProp(mac)
}
\arguments{
\item{snps}{data frame of SNP genotypes with first column called 'id', 
second column called 'stratum' and every two columns thereafter being 
SNP loci.}

\item{mac}{a vector of numerics giving the count of the minor allele, where 
0 = homozygous for reference allele, 1 = heterozygous, 2 = homozygous for
minor allele}
}
\value{
list of overall ('global') and by-strata ('pop') allele frequencies
}
\description{
Calculate overall and by-strata allele frequencies for input to 
  rmetasim.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
