% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGenotypes.R
\name{loadGenotypes}
\alias{loadGenotypes}
\title{Load Genotypes}
\usage{
loadGenotypes(label, scenario, replicate, folder = getwd())
}
\arguments{
\item{label}{label of simulation run to download.}

\item{scenario}{scenario number to retrieve.}

\item{replicate}{replicate number within the scenario to retrieve.}

\item{folder}{folder containing the folder of replicates. Default is current 
working directory.}
}
\value{
data frame of genotypes
}
\description{
Load a data frame of genotypes based on run label, scenario and 
  replicate number, and folder.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
