% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadScenarios.R
\name{loadScenarios}
\alias{loadScenarios}
\title{Load Scenarios}
\usage{
loadScenarios(label, folder = getwd())
}
\arguments{
\item{label}{label of simulation run.}

\item{folder}{folder containing the scenario data file. Default is current 
working directory.}
}
\value{
data frame of scenario specifications.
}
\description{
Load a data frame of scenario specifications for a simulation 
  run.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
