% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeEventSettings.R
\name{makeEventSettings}
\alias{makeEventSettings}
\title{Make Event Settings}
\usage{
makeEventSettings(dvgnc.time, num.pops)
}
\arguments{
\item{dvgnc.time}{number of generations since divergence.}

\item{num.pops}{number of populations.}
}
\value{
fscEvents object output from strataG::fscSettingsEvents.
}
\description{
Create event input for strataG::fscRun.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
