% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMigMat.R
\name{makeMigMat}
\alias{makeMigMat}
\title{Make Migration Matrix}
\usage{
makeMigMat(mig.rate, num.pops, type = c("island", "stepping.stone"))
}
\arguments{
\item{mig.rate}{migration rate specified as proportion of population 
migrating per generaton.}

\item{num.pops}{number of populations.}

\item{type}{of migration matrix structure. "island" = rate between all 
populations is the same. "stepping.stone" = migration only occurs between 
neighboring populations. Metapopulation is ring shaped, not a linear chain.}
}
\value{
matrix of pairwise migration rates.
}
\description{
Create island or stepping stone migration matrix given 
  migration rate and number of populations.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
