% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScenarios.R
\name{makeScenarios}
\alias{makeScenarios}
\title{Make Scenario Parameters}
\usage{
makeScenarios(
  num.pops,
  Ne,
  num.samples,
  mig.rate,
  mig.type,
  dvgnc.time,
  marker.type,
  mut.rate,
  num.loci,
  ploidy,
  rmetasim.ngen
)
}
\arguments{
\item{num.pops}{number of populations.}

\item{Ne}{effective population size.}

\item{num.samples}{number of samples to simulate. Must be <= Ne.}

\item{mig.rate}{migration rate specified as proportion of population 
migrating per generaton.}

\item{mig.type}{of migration matrix structure. "island" = rate between all 
populations is the same. "stepping.stone" = migration only occurs between 
neighboring populations. Metapopulation is ring shaped, not a linear chain.}

\item{dvgnc.time}{number of generations since divergence.}

\item{marker.type}{type of marker to simulate ("snp").}

\item{mut.rate}{mutation rate of marker in units of # of mutations per 
generation.}

\item{num.loci}{number of independent loci.}

\item{ploidy}{ploidy of marker. Set to 2 for diploid.}

\item{rmetasim.ngen}{number of generations to run Rmetasim for. Set to 
\code{0} to skip Rmetasim.}
}
\value{
data frame of scenario specifications.
}
\description{
Create data frame of scenario parameters.
}
\note{
Each parameter can be a single value or vector of values. Output will 
  be all unique combinations of values, with the following exceptions. 
  Scenarios wth \code{num.samples} > \code{Ne} are deleted. Scenarios with
  \code{num.pops = 1} will have \code{mig.rate} set to NA.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
