% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repCSVfiles.R
\name{repCSVfiles}
\alias{repCSVfiles}
\title{Get replicate .csv file names}
\usage{
repCSVfiles(folder)
}
\arguments{
\item{folder}{path of folder to search in.}
}
\description{
Return data frame of replicate file names, and extract
scenario and replicate numbers
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
