% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEBVsim.R
\name{runEBVsim}
\alias{runEBVsim}
\title{Run EBV Simulations}
\usage{
runEBVsim(params, num.cores, fsc.exec = "fsc26")
}
\arguments{
\item{params}{list of parameters for scenario runs. See \code{Details} for 
description of format.}

\item{num.cores}{number of cores to use.}

\item{fsc.exec}{name of fastsimcoal executable.}
}
\value{
vector of folders where scenario data and replicates were written.
}
\description{
Run a set of scenarios and replicates to generate genotypes.
}
\details{
\code{params} must be a list with the following elements:
\tabular{ll}{
  \code{label} \tab label for the run.\cr
  \code{scenarios} \tab data frame of scenario specifications generated by \code{\link{makeScenarios}}.\cr
  \code{genetics} \tab specification of genetic data to be simulated as output from \code{\link[strataG]{fscSettingsGenetics}}.\cr
  \code{ploidy} \tab ploidy of genetic data.\cr
  \code{num.rep} \tab number of replicates to run for each scenario.\cr
  \code{google.drive.id} \tab id of root Google Drive to download run from.\cr
 }
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
