% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFscSim.R
\name{runFscSim}
\alias{runFscSim}
\title{Run fastsimcoal}
\usage{
runFscSim(label, sc, genetics, ploidy, num.rep, use.wd)
}
\arguments{
\item{label}{label of simulation run.}

\item{sc}{a one row data frame of scenario parameters.}

\item{genetics}{specification of genetic data to be simulated as output from 
\code{\link[strataG]{fscSettingsGenetics}}.}

\item{ploidy}{ploidy of genetic data.}

\item{num.rep}{number of replicates to run for each scenario.}

\item{use.wd}{use working directory for fastsimcoal files?}
}
\value{
parameter output from `strataG::fscRun()`
}
\description{
Run a fastsimcoal simulation given scenario parameters
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
