% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRmetasim.R
\name{runRmetasim}
\alias{runRmetasim}
\title{Run rmetasim}
\usage{
runRmetasim(freqs, sc, num.gens = 2)
}
\arguments{
\item{freqs}{list of allele frequencies generated by \code{\link{calcFreqs}}.}

\item{sc}{a one row data frame of scenario parameters.}

\item{num.gens}{number of generations to run rmetasim.}
}
\value{
rmetasim landscape after `num.gens` generations of simulation
}
\description{
Run rmetasim simulation given initial allele frequencies 
  and scenario parameters.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
