% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeScenarios.R
\name{writeScenarios}
\alias{writeScenarios}
\title{Write Scenarios}
\usage{
writeScenarios(label, scenarios, google.drive.id = NULL)
}
\arguments{
\item{label}{label of simulation run.}

\item{scenarios}{data frame of scenario specifications generated by 
\code{\link{makeScenarios}}.}

\item{google.drive.id}{id of root Google Drive to copy data to.}
}
\value{
list of output folders.
}
\description{
Write scenario data and setup folders for writing 
  genotype files from replicates
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
