% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeReps.R
\name{analyzeReps}
\alias{analyzeReps}
\title{Analyze Replicates}
\usage{
analyzeReps(folder, scenarios = NULL, analyses = NULL, num.cores = 1)
}
\arguments{
\item{folder}{folder containing replicate .csv files.}

\item{scenarios}{vector of scenario numbers to analyze.}

\item{analyses}{named list of EBV analyses to run.}

\item{num.cores}{number of cores to use to split replicate analyses among.}
}
\value{
list of data frames of summary metrics
}
\description{
Conduct analysis of all replicates for a given EBV
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
