% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeReps.R
\name{analyzeReps}
\alias{analyzeReps}
\title{Analyze Replicates}
\usage{
analyzeReps(analysis, label)
}
\arguments{
\item{analysis}{character vector of EBV analysis to run (`ldNe`, `g2`, 
`het`, or `froh`).}

\item{label}{label of simulation run.}
}
\value{
data frame of summary metrics
}
\description{
Conduct analysis of all replicates for a given EBV
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
