% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScenarios.R
\name{makeScenarios}
\alias{makeScenarios}
\title{Make Scenario Parameters}
\usage{
makeScenarios(num.pops, Ne, num.samples, mig.rate, mig.type = c("island",
  "stepping.stone"), dvgnc.time)
}
\arguments{
\item{num.pops}{number of populations.}

\item{Ne}{effective population size.}

\item{num.samples}{number of samples to simulate. Must be <= Ne.}

\item{mig.rate}{migration rate specified as proportion of population 
migrating per generaton.}

\item{mig.type}{of migration matrix structure. "island" = rate between all 
populations is the same. "stepping.stone" = migration only occurs between 
neighboring populations. Metapopulation is ring shaped, not a linear chain.}

\item{dvgnc.time}{number of generations since divergence.}
}
\value{
data frame of scenario specifications.
}
\description{
Create data frame of scenario parameters.
}
\note{
Each parameter can be a single value or vector of values. Output will 
  be all unique combinations of values.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
