% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEBVsim.R
\name{runEBVsim}
\alias{runEBVsim}
\title{Run EBV Simulations}
\usage{
runEBVsim(
  label,
  scenarios,
  num.rep,
  google.drive.id = NULL,
  delete.fsc.files = TRUE,
  use.wd = FALSE,
  num.cores = 1,
  fsc.exec = "fsc26"
)
}
\arguments{
\item{label}{label used for output files and folders for the run.}

\item{scenarios}{data.frame where each row provides different parameters for 
each scenario. See \code{Details} for more information on its structure.}

\item{num.rep}{number of replicates to run for each scenario.}

\item{google.drive.id}{id of root Google Drive to save run to.}

\item{delete.fsc.files}{delete fastsimcoal files when the run is complete?}

\item{use.wd}{use the working directory to save fastsimcoal files? If 
\code{FALSE} (default), a temporary folder is used, which is deleted when 
the R session is closed.}

\item{num.cores}{number of cores to use. Each scenario is allocated to a 
separate core. If set to 1, then progress is output to the console.}

\item{fsc.exec}{name of fastsimcoal executable.}
}
\value{
vector of folders where scenario data and replicates were written.
}
\description{
Run a set of scenarios and replicates to generate genotypes.
}
\details{
The \code{scenarios} data.frame must have the following columns:
\tabular{ll}{
  \code{num.pops} \tab number of populations.\cr
  \code{Ne} \tab effective population size of each population.\cr
  \code{num.samples} \tab number of samples to use.\cr
  \code{mig.rate} \tab migration rate.\cr
  \code{mig.type} \tab type of migration matrix: "island" or "stepping.stone".\cr
  \code{dvgnc.time} \tab divergence time of populations.\cr
  \code{rmetasim.ngen} \tab number of generations to run \code{rmetasim} for.\cr
 }
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
