% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEBVsim.R
\name{runEBVsim}
\alias{runEBVsim}
\title{Run EBV Simulations}
\usage{
runEBVsim(label, scenarios, genetics, num.rep = 1, ploidy = 2,
  google.drive.id = NULL, run.rmetasim = TRUE)
}
\arguments{
\item{label}{label of simulation run to download.}

\item{scenarios}{data frame of scenario specifications generated by 
\code{\link{makeScenarios}}.}

\item{genetics}{specification of genetic data to be simulated as output from 
\code{\link[strataG]{fscSettingsGenetics}}.}

\item{num.rep}{number of replicates to run for each scenario.}

\item{ploidy}{ploidy of genetic data.}

\item{google.drive.id}{id of root Google Drive to download run from.}

\item{run.rmetasim}{logical determining whether or not to run rmetasim for 
two generations after fastsimcoal run.}
}
\value{
vector of folders where scenario data and replicates were written.
}
\description{
Run a set of scenarios and replicates to generate genotypes.
}
\author{
Eric Archer \email{eric.archer@noaa.gov}
}
